// packages/web/src/shared/stores/demo/types.ts
export interface DemoNominationsState {
  nominationLists: MockNominationLists
  possiblePlayers: PossibleNomination[]
  scorers: PersonSearchResult[]
}

/**
 * Validation slice state.
 */
export interface DemoValidationState {
  validatedGames: Record<string, ValidatedGameData>
  pendingScorers: Record<string, PendingScorerData>
}

/**
 * Combined state for all demo slices.
 */
export interface DemoState
  extends